#39 feat: Add depth2 navigation feature - Add depth2 redux file - Add re-comment component - Add cafe comment page layout component
