+++ src/components/organism/cafeWriteComment/index.tsx
@@ -1,0 +1,11 @@
+import { Box, Typography } from '@mui/material';
+
+const CafeWriteComment = () => {
+  return (
+    <Box>
+      <Typography>댓글 작성하기</Typography>
+      <Typography>댓글 작성하기</Typography>
+    </Box>
+  );
+};
+export default CafeWriteComment;
